+++ packages/datadog_flutter_plugin/example/ios/Tests/Helpers/ContractHelpers.swift
@@ -65,6 +65,7 @@ func testContracts(contracts: [Contract], plugin: FlutterPlugin, additionalArgum
         switch result {
         case .called(let value):
             let error = value as? FlutterError
+            XCTAssertNotEqual(value as? NSObject, FlutterMethodNotImplemented)
             XCTAssertNil(error, "\(contract.methodName) returned result \(String(describing: error)) on valid call")
 
         case .notCalled: