+++ packages/datadog_flutter_plugin/integration_test_app/lib/integration_scenarios/integration_scenarios_screen.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/material.dart';
 
 import '../auto_integration_scenarios/rum_auto_instrumentation_scenario.dart';
+import 'kiosk_integration_scenario.dart';
 import 'logging_scenario.dart';
 import 'rum_manual_error_reporting_scenario.dart';
 import 'rum_manual_instrumentation_scenario.dart';
@@ -42,6 +43,10 @@ class _IntegrationScenariosScreenState
       label: 'Auto RUM Scenario',
       navItem: RumAutoInstrumentationScenario.new,
     ),
+    ScenarioItem(
+      label: 'Kiosk RUM Scenario',
+      navItem: KioskIntegrationScenario.new,
+    )
   ];
 
   @override